HREPTWO-250: show workflow
Dialogs can use the ServiceReference and ServiceTracker classes to reference and track services.
This is serialization-safe.  Note that the WorkflowDescriptors, as used by the dialogs, now only work
because of the way Wicket serializes.  The last page is kept in memory, older ones are flushed.  So
when a dialog receives an ajax request, it is still in memory.  The original page can no longer use the
descriptor though, as it will be null on deserialization.



svn path=/hippo-ecm/trunk/; revision=11314

+++ repository/addon/frontend/src/main/java/org/hippoecm/frontend/plugins/standards/tabs/TabsPlugin.java
@@ -145,7 +145,9 @@ public class TabsPlugin extends RenderPlugin {
     void onSelect(Tab tabbie, AjaxRequestTarget target) {
         if (tabbie.renderer instanceof IDynamicService) {
             IDynamicService service = (IDynamicService) tabbie.renderer;
-            service.delete();
+            if (service.canDelete()) {
+                service.delete();
+            }
         }
     }
